Move table creation/modification classes into table package
