+++ sql/src/main/java/io/crate/execution/ddl/TransportSchemaUpdateAction.java
@@ -25,6 +25,8 @@ package io.crate.execution.ddl;
 import com.carrotsearch.hppc.cursors.ObjectObjectCursor;
 import io.crate.Constants;
 import io.crate.action.FutureActionListener;
+import io.crate.execution.ddl.SchemaUpdateRequest;
+import io.crate.execution.ddl.SchemaUpdateResponse;
 import io.crate.execution.support.ActionListeners;
 import io.crate.metadata.IndexParts;
 import io.crate.metadata.PartitionName;
