+++ sql/src/main/java/io/crate/execution/ddl/tables/CreateAnalyzerTask.java
@@ -20,7 +20,7 @@
  * agreement.
  */
 
-package io.crate.execution.ddl;
+package io.crate.execution.ddl.tables;
 
 import io.crate.data.Row;
 import io.crate.data.Row1;
