+++ sql/src/main/java/io/crate/execution/ddl/tables/DropTableTask.java
@@ -20,7 +20,7 @@
  * agreement.
  */
 
-package io.crate.execution.ddl;
+package io.crate.execution.ddl.tables;
 
 import io.crate.data.InMemoryBatchIterator;
 import io.crate.data.Row;
