+++ sql/src/main/java/io/crate/planner/node/ddl/CreateAnalyzerPlan.java
@@ -26,7 +26,7 @@ package io.crate.planner.node.ddl;
 import io.crate.expression.symbol.SelectSymbol;
 import io.crate.data.Row;
 import io.crate.data.RowConsumer;
-import io.crate.execution.ddl.CreateAnalyzerTask;
+import io.crate.execution.ddl.tables.CreateAnalyzerTask;
 import io.crate.planner.Plan;
 import io.crate.planner.PlannerContext;
 import io.crate.planner.DependencyCarrier;
