+++ sql/src/main/java/io/crate/planner/node/ddl/DropTablePlan.java
@@ -24,7 +24,7 @@ package io.crate.planner.node.ddl;
 import io.crate.expression.symbol.SelectSymbol;
 import io.crate.data.Row;
 import io.crate.data.RowConsumer;
-import io.crate.execution.ddl.DropTableTask;
+import io.crate.execution.ddl.tables.DropTableTask;
 import io.crate.metadata.doc.DocTableInfo;
 import io.crate.planner.DependencyCarrier;
 import io.crate.planner.Plan;
